+++ example/lib/usage.dart
@@ -1,9 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:internationalization/internationalization.dart';
 
+import 'infra/res/internationalization.gen.dart';
+import 'next_page.dart';
+
 class Usage extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
+    Internationalization.of(context);
+
     return Scaffold(
       body: Row(
         mainAxisAlignment: MainAxisAlignment.center,
@@ -11,15 +16,44 @@ class Usage extends StatelessWidget {
           Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: <Widget>[
-              Text(Strings.of(context).valueOf("simple_string")),
-              Text(Strings.of(context).valueOf("interpolation_string", args: ["IM", "HERE"])),
-              Text(Strings.of(context).pluralOf("simple_plurals", 0)),
-              Text(Strings.of(context).pluralOf("simple_plurals", 1)),
-              Text(Strings.of(context).pluralOf("simple_plurals", 132548132)),
-              Text(Strings.of(context).pluralOf("interpolation_plurals", 0, args: ["IM", "HERE"])),
-              Text(Strings.of(context).pluralOf("interpolation_plurals", 1, args: ["IM", "HERE"])),
-              Text(Strings.of(context).pluralOf("interpolation_plurals", 132548132,args: ["IM", "HERE"])),
-              Text(Strings.of(context).valueOf("interpolation_string_with_named_args", args: ["IM"], namedArgs: {"named_arg_key": "Named arg"})),
+              Text("simple_string".translate()),
+              Text(Intl.usage.interpolationString(args: ["( ͡° ͜ʖ ͡°)"])),
+              Text(
+                Intl.usage.interpolationStringWithNamedArgs(
+                  namedArgs: {"named_arg_key": "( ͡° ͜ʖ ͡°)"},
+                ),
+              ),
+              Text(Intl.usage.simplePlurals(pluralValue: 0)),
+              Text(Intl.usage.simplePlurals(pluralValue: 1)),
+              Text(Intl.usage.simplePlurals(pluralValue: 123456789)),
+              Text(
+                Intl.usage.interpolationPlurals(
+                  pluralValue: 0,
+                  args: ["( ͡° ͜ʖ ͡°)"],
+                ),
+              ),
+              Text(
+                Intl.usage.interpolationPlurals(
+                  pluralValue: 1,
+                  args: ["( ͡° ͜ʖ ͡°)"],
+                ),
+              ),
+              Text(
+                Intl.usage.interpolationPlurals(
+                  pluralValue: 123456789,
+                  args: ["123456789"],
+                ),
+              ),
+              RaisedButton(
+                child: Text(Intl.usage.nextPageBtn()),
+                onPressed: () {
+                  return Navigator.of(context).push(
+                    MaterialPageRoute(
+                      builder: (_) => NextPage(),
+                    ),
+                  );
+                },
+              ),
             ],
           ),
         ],